+++ app/(tabs)/index.tsx
@@ -1,24 +1,12 @@
-import { useState } from "react";
+import { Provider } from "react-redux";
+import { store } from "@/store";
+
 import { View, Dimensions, StyleSheet } from "react-native";
 
-import { ImageContentFit } from "expo-image";
-import * as ImagePicker from "expo-image-picker";
-import * as MediaLibrary from "expo-media-library";
-
-import Ionicons from "@expo/vector-icons/Ionicons";
-
-import ImageCropper from "../ImageCropper";
-
 import ImageViewer from "@/components/create/ImageViewer";
-import ImageSelectButton from "@/components/create/ImageSelectButton";
 import ActionsBar from "@/components/create/ActionsBar";
-import CropGuide from "@/components/create/CropGuide";
-import Button from "@/components/shared/Button";
 
 export default function Index() {
-  const [image, setImage] = useState<string | null>(null);
-  const [status, requestPermission] = MediaLibrary.usePermissions();
-
   const SocialMedia = [
     {
       name: "Instagram",
@@ -28,10 +16,6 @@ export default function Index() {
       },
     },
   ];
-
-  if (status === null) {
-    requestPermission();
-  }
 
   const targetSocialMedia = SocialMedia.find(
     (media) => media.name === "Instagram"
@@ -43,86 +27,22 @@ export default function Index() {
 
   const { width, height } = socialMediaImageRatio;
 
-  const quality = "720p";
-
   const viewerStyle = {
     width: Dimensions.get("window").width,
     height: (Dimensions.get("window").width / width) * height,
   };
 
-  const pickImageAsync = async () => {
-    let result = await ImagePicker.launchImageLibraryAsync({
-      quality: 1,
-    });
-
-    if (!result.canceled) {
-      setImage(result.assets[0].uri);
-    }
-  };
-
-  const resetImage = () => {
-    setImage(null);
-  };
-
-  const ImageCropperOptions = {
-    rows: 1,
-    cols: 3,
-    compress: 1,
-  };
-
-  const runImageCropper = async () => {
-    if (image) {
-      const imageCropper = new ImageCropper(image, ImageCropperOptions);
-
-      return await imageCropper.run();
-    }
-    return [];
-  };
-
-  const saveImageAsync = async () => {
-    try {
-      const croppedImages = await runImageCropper();
-
-      for (const croppedImage of croppedImages) {
-        await MediaLibrary.saveToLibraryAsync(croppedImage.uri);
-      }
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
   return (
-    <View style={styles.container}>
-      {image ? (
+    <Provider store={store}>
+      <View style={styles.container}>
         <View style={[styles.imageContainer, viewerStyle]}>
-          <ImageViewer imgSource={image} />
+          <ImageViewer />
         </View>
-      ) : (
-        <View style={[styles.imageContainer, viewerStyle]}>
-          <ImageSelectButton onPress={pickImageAsync} />
+        <View style={styles.chinContainer}>
+          <ActionsBar />
         </View>
-      )}
-      <View style={styles.chinContainer}>
-        <ActionsBar>
-          <Button
-            label={<Ionicons name="refresh-outline" size={24} />}
-            disabled={!image}
-            onPress={resetImage}
-          />
-          <Button label={`${width}:${height}`} disabled={true} />
-          <Button
-            label={`${ImageCropperOptions.rows}x${ImageCropperOptions.cols}`}
-            disabled={true}
-          />
-          <Button label={quality} disabled={true} />
-          <Button
-            label={<Ionicons name="download-outline" size={24} />}
-            disabled={!image}
-            onPress={saveImageAsync}
-          />
-        </ActionsBar>
       </View>
-    </View>
+    </Provider>
   );
 }